Added Composite reporter inside rules_scala
