+++ test_expect_failure/diagnostics_reporter/ErrorAndWarningFile.scala
@@ -1,0 +1,7 @@
+import java.util.ArrayList
+
+object ErrorAndWarningFile {
+  def main(args: Array[String]): Unit = {
+    printn("this is not a print")
+  }
+}
