+++ test_expect_failure/diagnostics_reporter/InfoFile.scala
@@ -1,0 +1,5 @@
+object InfoFile {
+  def main(args: Array[String]): Unit = {
+    println("this is not a print")
+  }
+}
