+++ test_expect_failure/diagnostics_reporter/SimpleFile.scala
@@ -1,0 +1,7 @@
+object SimpleFile {
+  val value = "someVal"
+
+  def main(args: Array[String]): Unit = {
+    println(value
+  }
+}
