+++ test_expect_failure/diagnostics_reporter/TwoErrorsFile.scala
@@ -1,0 +1,8 @@
+object SimpleFile {
+  val value = "someVal"
+
+  def main(args: Array[String]): Unit = {
+    printn(value)
+    prinf(value)
+  }
+}
