+++ test_expect_failure/diagnostics_reporter/WarningFile.scala
@@ -1,0 +1,7 @@
+import java.util.ArrayList
+
+object WarningFile {
+  def main(args: Array[String]): Unit = {
+    println("this is a print")
+  }
+}
